created df to show lists activity and integrated with progress.

+++ dbt_project/analyses/completed_counts.sql
@@ -30,7 +30,7 @@ task_level as (
 select 
 fld_folder_name,
 l_list_name,
-sum(cnt) as tasks_completed,
+sum(cnt)::int as tasks_completed,
 max_day_completed_timestamp,
  day,month,year
